Fix Hungarian footer to match the official style
When we using the Hungarian language option, the left footer text is messed up. It shows "{curr}/{tot} Oldal - {name} Önéletrajza" (translation: "{curr}/{tot} Page - Curriculum vitae of {name}") and not the build date. The other languages and the official online editor outputs a PDF with a date in a footer.